requirements.txt
numpy==1.20.1
scipy==1.6.2
matplotlib==3.4.2
protobuf==3.17.3
PyYAML==5.4.1
